added base entity and commonbase repository
