+++ softgroup/basis/common-dao/src/main/java/com/softgroup/common/dao/api/CommonBaseRepository.java
@@ -1,0 +1,10 @@
+package com.softgroup.common.dao.api;
+
+import org.springframework.data.repository.PagingAndSortingRepository;
+
+/**
+ * Created by yuriy on 23.03.17.
+ */
+public interface CommonBaseRepository<T>
+        extends PagingAndSortingRepository<T, String> {
+}
