+++ softgroup/basis/common-dao/src/main/java/com/softgroup/common/dao/entities/ProfileSettingsEntity.java
@@ -1,5 +1,6 @@
 package com.softgroup.common.dao.entities;
 
+import com.softgroup.common.dao.api.BaseEntity;
 import org.hibernate.annotations.GenericGenerator;
 
 import javax.persistence.*;
@@ -10,29 +11,18 @@ import java.io.Serializable;
  */
 @Entity
 @Table(name = "profile_settings")
-public class ProfileSettingsEntity implements Serializable {
+public class ProfileSettingsEntity
+        extends BaseEntity
+        implements Serializable {
 
     private static final long serialVersionUID = 2645460488213358603L;
-
-    @Id
-    @Column(name = "id")
-    @GeneratedValue(generator = "system-uuid")
-    @GenericGenerator(name = "system-uuid", strategy = "uuid")
-    private String id;
 
     @Column(name = "settings_data")
     private String settingsData;
 
     @ManyToOne(fetch = FetchType.EAGER)
+    @JoinColumn(name = "profile_id")
     private ProfileEntity profile;
-
-    public String getId() {
-        return id;
-    }
-
-    public void setId(String id) {
-        this.id = id;
-    }
 
     public String getSettingsData() {
         return settingsData;
@@ -57,13 +47,13 @@ public class ProfileSettingsEntity implements Serializable {
 
         ProfileSettingsEntity that = (ProfileSettingsEntity) o;
 
-        if (!id.equals(that.id)) return false;
+        if (!getId().equals(that.getId())) return false;
         return settingsData.equals(that.settingsData);
     }
 
     @Override
     public int hashCode() {
-        int result = id.hashCode();
+        int result = getId().hashCode();
         result = 31 * result + settingsData.hashCode();
         return result;
     }
