+++ softgroup/basis/common-dao/src/main/java/com/softgroup/common/dao/repository/ProfileRepository.java
@@ -1,5 +1,6 @@
 package com.softgroup.common.dao.repository;
 
+import com.softgroup.common.dao.api.CommonBaseRepository;
 import com.softgroup.common.dao.entities.ProfileEntity;
 import org.springframework.data.jpa.repository.Query;
 import org.springframework.data.repository.PagingAndSortingRepository;
@@ -11,7 +12,7 @@ import java.util.List;
  * Created by yuriy on 21.03.17.
  */
 public interface ProfileRepository
-        extends PagingAndSortingRepository<ProfileEntity, String> {
+        extends CommonBaseRepository<ProfileEntity> {
 
     List<ProfileEntity> findByName(String s);
