+++ softgroup/basis/common-dao/src/main/java/com/softgroup/common/dao/repository/ProfileSettingsRepository.java
@@ -1,5 +1,6 @@
 package com.softgroup.common.dao.repository;
 
+import com.softgroup.common.dao.api.CommonBaseRepository;
 import com.softgroup.common.dao.entities.ProfileSettingsEntity;
 import org.springframework.data.repository.PagingAndSortingRepository;
 
@@ -7,5 +8,5 @@ import org.springframework.data.repository.PagingAndSortingRepository;
  * Created by yuriy on 21.03.17.
  */
 public interface ProfileSettingsRepository
-        extends PagingAndSortingRepository<ProfileSettingsEntity, String> {
+        extends CommonBaseRepository<ProfileSettingsEntity> {
 }
