+++ softgroup/basis/common-dao/src/main/java/com/softgroup/common/dao/service/ProfileService.java
@@ -1,10 +1,8 @@
 package com.softgroup.common.dao.service;
 
 import com.softgroup.common.dao.api.AbstractCrudService;
-import com.softgroup.common.dao.api.CrudService;
 import com.softgroup.common.dao.entities.ProfileEntity;
 import com.softgroup.common.dao.repository.ProfileRepository;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 /**
@@ -12,6 +10,6 @@ import org.springframework.stereotype.Service;
  */
 @Service
 public class ProfileService
-        extends CrudService<ProfileEntity, ProfileRepository> {
+        extends AbstractCrudService<ProfileEntity, ProfileRepository> {
 
 }
