+++ softgroup/basis/common-dao/src/main/java/com/softgroup/common/dao/service/ProfileSettingsService.java
@@ -1,7 +1,6 @@
 package com.softgroup.common.dao.service;
 
 import com.softgroup.common.dao.api.AbstractCrudService;
-import com.softgroup.common.dao.api.CrudService;
 import com.softgroup.common.dao.entities.ProfileSettingsEntity;
 import com.softgroup.common.dao.repository.ProfileSettingsRepository;
 
@@ -9,5 +8,5 @@ import com.softgroup.common.dao.repository.ProfileSettingsRepository;
  * Created by yuriy on 21.03.17.
  */
 public class ProfileSettingsService
-        extends CrudService<ProfileSettingsEntity, ProfileSettingsRepository> {
+        extends AbstractCrudService<ProfileSettingsEntity, ProfileSettingsRepository> {
 }
